+++ src/components/Leaderboard/Leaderboard.scss
